Implement Google reCAPTCHA in contact form
The HTML and PHP files have been updated to integrate Google's reCAPTCHA service into the contact form. Changes include the insertion of the reCAPTCHA script into the HTML head and adjustments to the PHP file to verify the reCAPTCHA response. This will provide an additional layer of security against spam and abuse on the website.